services/orchestrator/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
sqlalchemy
psycopg2-binary
langgraph
langchain-openai
langchain-core
httpx
redis
jinja2
python-multipart
docker